asignacion instrumento

+++ src/main/java/LengBD/domain/AsignacionListadoDTO.java
@@ -6,12 +6,12 @@ import java.time.LocalDate;
 @Data
 public class AsignacionListadoDTO {
     private Integer idAsignacion;
-    private String nombreIntegrante;    // para la tabla
-    private String nombreInstrumento;   // para la tabla
+    private String nombreIntegrante;    
+    private String nombreInstrumento;   
     private LocalDate fechaInicio;
     private LocalDate fechaFinal;
     private String motivo;
-    private Integer cedula;             // para el editar
-    private Integer idInstrumento;      // para el editar
-    private Integer idEstado;           // para el editar
+    private Long cedula;             
+    private Integer idInstrumento;      
+    private Integer idEstado;           
 }
